refactor(9): consolidar src/ dentro de jobhunter/
- src/cv_builder.py -> jobhunter/cv/builder.py
- src/cv_templates/ -> jobhunter/cv/templates/
- src/offer_utils.py -> jobhunter/offers.py
- Actualizados todos los imports internos de cv_templates y tests
- Agregado tests/test_offers.py cubre al modulo con casos existentes
- src/ eliminado completamente
- 103 tests pasan
