+++ jobhunter/cv/templates/compact.py
@@ -7,7 +7,7 @@ from reportlab.lib.colors import HexColor
 from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
 from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, HRFlowable, Table, TableStyle
 from reportlab.lib.enums import TA_LEFT, TA_RIGHT, TA_JUSTIFY
-from src.cv_builder import _safe, SECTION_LABELS, safe_header_name
+from jobhunter.cv.builder import _safe, SECTION_LABELS, safe_header_name
 
 PRIMARY = HexColor("#0d47a1")
 TEXT_DARK = HexColor("#212121")
