+++ jobhunter/cv/templates/modern.py
@@ -7,7 +7,7 @@ from reportlab.lib.colors import HexColor
 from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
 from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, HRFlowable
 from reportlab.lib.enums import TA_LEFT, TA_JUSTIFY
-from src.cv_builder import _safe, SECTION_LABELS, safe_header_name
+from jobhunter.cv.builder import _safe, SECTION_LABELS, safe_header_name
 
 PRIMARY = HexColor("#1a1a2e")
 ACCENT = HexColor("#0f3460")
