+++ jobhunter/offers.py
